Rename *FieldModelContext to IndexSchemaField*Context
This should be more consistent with both IndexSchemaElement and
ElasticsearchIndexSchemaNode.
