+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/document/model/ElasticsearchIndexSchemaFieldTypedContext.java
@@ -4,15 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.document.model;
+package org.hibernate.search.v6poc.backend.elasticsearch.document.model;
 
-import org.hibernate.search.v6poc.backend.document.IndexFieldAccessor;
+import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldTypedContext;
+
 
 /**
  * @author Yoann Rodiere
  */
-public interface TerminalFieldModelContext<T> {
-
-	IndexFieldAccessor<T> createAccessor();
+public interface ElasticsearchIndexSchemaFieldTypedContext<T> extends IndexSchemaFieldTypedContext<T> {
 
 }
