+++ engine/src/main/java/org/hibernate/search/v6poc/backend/document/model/IndexSchemaFieldTerminalContext.java
@@ -4,14 +4,15 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.elasticsearch.document.model;
+package org.hibernate.search.v6poc.backend.document.model;
 
-import org.hibernate.search.v6poc.backend.document.model.TypedFieldModelContext;
-
+import org.hibernate.search.v6poc.backend.document.IndexFieldAccessor;
 
 /**
  * @author Yoann Rodiere
  */
-public interface ElasticsearchTypedFieldModelContext<T> extends TypedFieldModelContext<T> {
+public interface IndexSchemaFieldTerminalContext<T> {
+
+	IndexFieldAccessor<T> createAccessor();
 
 }
