+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/bridge/builtin/impl/DefaultIntegerFunctionBridge.java
@@ -6,14 +6,14 @@
  */
 package org.hibernate.search.v6poc.entity.pojo.bridge.builtin.impl;
 
-import org.hibernate.search.v6poc.backend.document.model.FieldModelContext;
-import org.hibernate.search.v6poc.backend.document.model.TypedFieldModelContext;
+import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldContext;
+import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldTypedContext;
 import org.hibernate.search.v6poc.entity.pojo.bridge.FunctionBridge;
 
 public final class DefaultIntegerFunctionBridge implements FunctionBridge<Integer, Integer> {
 
 	@Override
-	public TypedFieldModelContext<Integer> bind(FieldModelContext context) {
+	public IndexSchemaFieldTypedContext<Integer> bind(IndexSchemaFieldContext context) {
 		return context.asInteger();
 	}
