+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/bridge/builtin/impl/DefaultLocalDateFunctionBridge.java
@@ -8,14 +8,14 @@ package org.hibernate.search.v6poc.entity.pojo.bridge.builtin.impl;
 
 import java.time.LocalDate;
 
-import org.hibernate.search.v6poc.backend.document.model.FieldModelContext;
-import org.hibernate.search.v6poc.backend.document.model.TypedFieldModelContext;
+import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldContext;
+import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldTypedContext;
 import org.hibernate.search.v6poc.entity.pojo.bridge.FunctionBridge;
 
 public final class DefaultLocalDateFunctionBridge implements FunctionBridge<LocalDate, LocalDate> {
 
 	@Override
-	public TypedFieldModelContext<LocalDate> bind(FieldModelContext context) {
+	public IndexSchemaFieldTypedContext<LocalDate> bind(IndexSchemaFieldContext context) {
 		return context.asLocalDate();
 	}
